refactor: SearchType을 게시판 공통 패키지로 이동
