+++ src/main/java/inha/gdgoc/domain/board/event/repository/EventBoardRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import inha.gdgoc.domain.board.event.entity.EventBoard;
 import inha.gdgoc.domain.board.event.entity.QEventBoard;
-import inha.gdgoc.domain.board.event.enums.SearchType;
+import inha.gdgoc.domain.board.common.enums.SearchType;
 import inha.gdgoc.domain.user.enums.TeamType;
 import inha.gdgoc.domain.user.enums.UserRole;
 import java.util.List;
